Update project structure to reflect GUI classes
Now that we have a GUI, we should re-organize the classes appropriately.

Inside "root" package `duke`, let's
- Remove text-based UI methods from Duke.java and put them in a new
class called TextDuke
- Rename the text UI class from `Ui` to `TextUi` for clear distinction
with the new GUI
- Put controller class `MainWindow` and `DialogBox` inside package `ui`
- Group packages `commands`, `parser`, `tasks` under a new package `logic`
to bundle together the inner workings of Duke
- Update unit test classes accordingly
