+++ src/main/java/duke/logic/commands/ExitCommand.java
@@ -1,4 +1,4 @@
-package duke.commands;
+package duke.logic.commands;
 
 /**
  * Exits the program.
