+++ src/main/java/duke/logic/tasks/Task.java
@@ -1,4 +1,4 @@
-package duke.tasks;
+package duke.logic.tasks;
 
 /**
  * Represents a task.
