+++ src/main/java/duke/ui/MainWindow.java
@@ -1,6 +1,8 @@
-package duke;
+package duke.ui;
 
+import duke.logic.Duke;
 import javafx.fxml.FXML;
+import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
@@ -9,7 +11,7 @@ import javafx.scene.layout.VBox;
 
 /**
  * Controller for MainWindow. Provides the layout for the other controls.
- * Directs the flow and interaction between UI elements + their callbacks and Duke i.e. the program's logic.
+ * Directs the flow and interaction between UI elements and the program's logic.
  */
 public class MainWindow extends AnchorPane {
     @FXML
@@ -18,9 +20,10 @@ public class MainWindow extends AnchorPane {
     private VBox dialogContainer;
     @FXML
     private TextField userInput;
-    // @FXML
-    // private Button sendButton;
+    @FXML
+    private Button sendButton;
 
+    /** The program's logic component */
     private Duke duke;
 
     private final Image userImage = new Image(this.getClass().getResourceAsStream("/images/BreadDoge.png"));
@@ -34,12 +37,13 @@ public class MainWindow extends AnchorPane {
     public void setUpDuke(Duke d) {
         duke = d;
         String initMessage = "Hello! What can I do for you?";
-        dialogContainer.getChildren().add(DialogBox.getDukeDialog(initMessage, dukeImage));
+        DialogBox initDialog = DialogBox.getDukeDialog(initMessage, dukeImage);
+        dialogContainer.getChildren().add(initDialog);
     }
 
     /**
-     * Creates two dialog boxes, one echoing user input and the other containing Duke's reply and then appends them to
-     * the dialog container. Clears the user input after processing.
+     * Creates two dialog boxes, one echoing user input and the other containing Duke's reply
+     * then appends them to the dialog container. Clears the user input after processing.
      */
     @FXML
     private void handleUserInput() {
