+++ src/test/java/duke/DukeTest.java
@@ -3,7 +3,7 @@ package duke;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
-
+// todo Delete this dummy class
 public class DukeTest {
     @Test
     public void dummyTest() {
