+++ src/test/java/duke/logic/tasks/EventTest.java
@@ -1,4 +1,4 @@
-package duke.tasks;
+package duke.logic.tasks;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
