+++ src/test/java/duke/logic/tasks/TaskListTest.java
@@ -1,4 +1,4 @@
-package duke.tasks;
+package duke.logic.tasks;
 
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
